Se ha implementado una forma de hacer consultas en base a un request {ip:"", direccionDTO:{}, idPeticion:x} Sin embargo se está presentando una incongruencia en la base de datos porque algunas variables no son capaces de soportar números tan grandes, por lo que se están haciendo las modificaciones necesarias.

+++ Publisher/src/main/java/com/telefonia/Entities/CoberturaColonia.java
@@ -31,7 +31,7 @@ public class CoberturaColonia {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cobertura_id")
-    private Short coberturaId;
+    private Integer coberturaId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "colonia_id", nullable = false, unique = true)
