+++ Publisher/src/main/java/com/telefonia/Entities/CoberturaEstado.java
@@ -31,7 +31,7 @@ public class CoberturaEstado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cobertura_id")
-    private Short coberturaId;
+    private Integer coberturaId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "estado_id", nullable = false, unique = true)
