+++ Publisher/src/main/java/com/telefonia/Entities/Servicio.java
@@ -30,7 +30,7 @@ public class Servicio {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "servicio_id")
-    private Short servicioId;
+    private Integer servicioId;
 
     @Column(name = "nombre", length = 100, nullable = false)
     private String nombre;
