+++ Publisher/src/main/java/com/telefonia/repository/ServicioRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.telefonia.Entities.Servicio;
 
 @Repository
-public interface ServicioRepository extends CrudRepository<Servicio, Short> {
+public interface ServicioRepository extends CrudRepository<Servicio, Integer> {
     // AQUÍ SE ESCRIBEN LOS QUERYS PERSONALIZADOS
     // ES UNA INTERFACE POR LO QUE NO SE IMPLEMENTAN LOS MÉTODOS
 }
